requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# AI/ML Libraries
langchain==0.1.0
langchain-openai==0.0.2
openai==1.6.1
sentence-transformers==2.5.1
huggingface-hub==0.20.3
transformers==4.36.0
torch==2.1.0

# Vector Database
faiss-cpu==1.7.4
chromadb==0.4.22
qdrant-client==1.7.0

# Evaluation
ragas==0.1.1
datasets==2.15.0

# Document Processing
PyPDF2==3.0.1
python-docx==1.1.0
beautifulsoup4==4.12.2
requests==2.31.0

# Utilities
numpy==1.24.3
pandas==2.1.3
tiktoken==0.5.2
aiohttp==3.9.1

# Deployment
gunicorn==21.2.0
python-multipart==0.0.6
